Move rebinding ui into a separate MonoBehaviour
