+++ Assets/TanksDemo/Assets/Scripts/UI/ControlRebindingUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69252387899be4887b73fe4f7dc4626a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
